refactor: улучшена обработка закрытия ресурсов и исправлены тесты
- Обновлены функции для закрытия HTTP-ответов, чтобы избежать игнорирования ошибок.
- Исправлены тесты для корректной обработки ошибок и улучшения читаемости кода.
- Изменены типы полей в структурах данных для повышения типобезопасности.
- Упрощены вызовы функций, связанные с обработкой ответов и удалением временных файлов.

+++ internal/service/gitsync/service.go
@@ -193,7 +193,7 @@ func (s *GitSyncService) syncCampus(ctx context.Context, campus db.Campuse, path
 
 		// Upsert Club
 		_, err = s.queries.UpsertClub(ctx, db.UpsertClubParams{
-			ID:           int32(c.ID),
+			ID:           int16(c.ID),
 			CampusID:     campus.ID,
 			LeaderLogin:  toText(c.LeaderLogin),
 			Name:         c.Name,
